requirements.txt
# ⚡ Telegram Bot API library
# Эта библиотека позволяет боту общаться с Telegram
# Версия 20.7 - проверенная и стабильная
python-telegram-bot==20.7

# 🔐 Для безопасного хранения токена
# Позволяет читать токен из файла .env
# Чтобы не светить его в коде
python-dotenv==1.0.0

# 🚀 Для асинхронных операций
# asyncio - уже встроена в Python

# 📝 Для логирования
# logging - уже встроена в Python

# 💾 Для работы с файловой системой
# os - уже встроена в Python